Move Authentification controller in a appropriate directory
