+++ Tools/Route/Authentification.php
@@ -1,5 +1,8 @@
 <?php declare(strict_types = 1);
+
+use LibertAPI\Tools\Controllers\AuthentificationController as Auth;
+
 /*
  * Doit être importé après la création de $app. Ne créé rien.
  */
-$app->get('/authentification', ['controller', 'get'])->setName('authentification');
+$app->get('/authentification', [Auth::class, 'get'])->setName('authentification');
